feature : refresh Cookies value fix
