+++ Auth-Server/src/main/java/com/dju/gdsc/domain/oauth/handler/OAuth2AuthenticationSuccessHandler.java
@@ -4,8 +4,8 @@ package com.dju.gdsc.domain.oauth.handler;
 import com.dju.gdsc.domain.member.model.RoleType;
 import com.dju.gdsc.domain.oauth.entity.ProviderType;
 import com.dju.gdsc.domain.oauth.entity.UserRefreshToken;
-import com.dju.gdsc.domain.oauth.info.OAuth2UserInfo;
-import com.dju.gdsc.domain.oauth.info.OAuth2UserInfoFactory;
+import com.dju.gdsc.domain.oauth.handler.info.OAuth2UserInfo;
+import com.dju.gdsc.domain.oauth.handler.info.OAuth2UserInfoFactory;
 import com.dju.gdsc.domain.oauth.repository.OAuth2AuthorizationRequestBasedOnCookieRepository;
 import com.dju.gdsc.domain.oauth.repository.UserRefreshTokenRepository;
 import com.dju.gdsc.domain.oauth.token.AuthToken;
@@ -29,11 +29,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.net.URI;
-import java.text.SimpleDateFormat;
 import java.util.Collection;
 import java.util.Date;
 import java.util.Optional;
-import java.util.TimeZone;
 
 import static com.dju.gdsc.domain.oauth.repository.OAuth2AuthorizationRequestBasedOnCookieRepository.REDIRECT_URI_PARAM_COOKIE_NAME;
 import static org.springframework.security.oauth2.core.endpoint.OAuth2ParameterNames.REFRESH_TOKEN;
