+++ Auth-Server/src/main/java/com/dju/gdsc/domain/oauth/handler/info/OAuth2UserInfo.java
@@ -1,0 +1,27 @@
+package com.dju.gdsc.domain.oauth.handler.info;
+
+import com.dju.gdsc.domain.oauth.entity.ProviderType;
+
+import java.util.Map;
+
+public abstract class OAuth2UserInfo {
+    protected Map<String, Object> attributes;
+
+    public OAuth2UserInfo(Map<String, Object> attributes) {
+        this.attributes = attributes;
+    }
+
+    public Map<String, Object> getAttributes() {
+        return attributes;
+    }
+
+    public abstract String getId();
+
+    public abstract String getName();
+
+    public abstract String getEmail();
+
+    public abstract String getImageUrl();
+
+    public abstract ProviderType getProvider();
+}
