+++ Auth-Server/src/main/java/com/dju/gdsc/domain/oauth/handler/info/OAuth2UserInfoFactory.java
@@ -1,0 +1,18 @@
+package com.dju.gdsc.domain.oauth.handler.info;
+
+
+
+import com.dju.gdsc.domain.oauth.entity.ProviderType;
+import com.dju.gdsc.domain.oauth.handler.info.impl.GoogleOAuth2UserInfo;
+
+import java.util.Map;
+
+public class OAuth2UserInfoFactory {
+    public static OAuth2UserInfo getOAuth2UserInfo(ProviderType providerType, Map<String, Object> attributes) {
+        switch (providerType) {
+            case GOOGLE: return new GoogleOAuth2UserInfo(attributes);
+
+            default: throw new IllegalArgumentException("Invalid Provider Type.");
+        }
+    }
+}
