+++ Auth-Server/src/main/java/com/dju/gdsc/domain/oauth/infrastructure/OauthRequester.java
@@ -1,7 +1,7 @@
 package com.dju.gdsc.domain.oauth.infrastructure;
 
 import com.dju.gdsc.domain.oauth.entity.ProviderType;
-import com.dju.gdsc.domain.oauth.info.OAuth2UserInfo;
+import com.dju.gdsc.domain.oauth.handler.info.OAuth2UserInfo;
 
 public interface OauthRequester {
     boolean supports(ProviderType provider);
